Fix bugs involving elimination and candidate set "methods".
- Handle eliminated candidates for the following methods:
	Condorcet, maximum element sets, strongly dominating set
- Fix bug in Condorcet when candidate 0 has been eliminated.

+++ src/singlewinner/sets/max_elements/det_sets.h
@@ -4,7 +4,7 @@
 // inside, and this set is the minimal for which that is true, the set generator
 // can be modeled by extending this maximal elements class.
 
-// The set-determining methods are classed as "election methods", but they 
+// The set-determining methods are classed as "election methods", but they
 // produce ties so often that they really just elect sets.
 
 #ifndef _SET_MAX_ELEMS
@@ -16,14 +16,14 @@
 class det_sets_relation {
 
 	protected:
-		virtual bool relation(const abstract_condmat & input, int a, 
+		virtual bool relation(const abstract_condmat & input, int a,
 				int b, const vector<bool> & hopefuls) const = 0;
 
 		// Is this - to have nested sets - required? Make it an
 		// option when invoking these set methods.
 
-		ordering nested_sets(const abstract_condmat & input, 
-				const vector<bool> & hopefuls, int limit) const;
+		ordering nested_sets(const abstract_condmat & input,
+				const vector<bool> & hopefuls, size_t limit) const;
 		ordering nested_sets(const abstract_condmat & input,
 				const vector<bool> & hopefuls) const {
 			return(nested_sets(input, hopefuls,
